Add files via upload
This project is completed with help of Computer vision and Convolutional Neural Network. The project full name is Real-time Facial Emotion Recognition System for Learning Engagement and it provide real time emotion of the students while the online class. The need for this project is important because due to COVID-19 pandemic mostly classes were online and some students were mocking them. So to overcome this problem and increase the productivity of an online lectures this model has been made.
I've used FER-2013 data set for this project and using CNN and frontal face haar cascade model of recognition for this project. We are successfully achieved the accuracy of 71% in training and 73 % in testing the model.